A bunch of format string fixes and some casts to force formats
No objections from Fred Kiefer ;)



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@38569 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSDisplayServer.m
@@ -1112,7 +1112,7 @@ GSCurrentServer(void)
 	{
 	  NSEvent *event = events[i];
 
-          NSLog(@"index %d %@", i, event);
+          NSLog(@"index %lu %@", (unsigned long) i, event);
         }
     }
   else
